Changing names
Jeff, I realized that It is better to change the names of all the files just to avoid confusion. i used your PS2 answer key as a template to get my Replication pdf and I forgot to change names when i commited. Now I changed all names on the file to "Replication_answer" just to be clear. info is the same tough, is just the name of the files